reorder release 0
Flow wise it feels more smooth to install the gem before opening the database. Since we start off at command line, we can install and then go to the sql prompt to open the database. I just opened another tab to get a prompt, but moving the steps to have install the gem first would eliminate the need to open another tab, or worse, if someone quit sql to go back to a command prompt and repeat the prior instruction. What do you think?